requirements.txt
# Core dependencies for CSV to JSON converter
pandas>=1.5.0
PyYAML>=6.0
requests>=2.28.0

# Google APIs dependencies
gspread>=5.7.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Additional utilities
python-dateutil>=2.8.2
urllib3>=1.26.0
certifi>=2022.0.0

# Optional: For better CSV handling
chardet>=5.1.0

# Optional: For progress bars and enhanced logging
tqdm>=4.64.0
colorlog>=6.7.0